requirements-ai.txt
# AI Enhancement Dependencies for Clearance Face Search System
# Install with: pip install -r requirements-ai.txt

# Core AI/ML Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
imageio>=2.31.0
scikit-image>=0.21.0

# Advanced Face Recognition
deepface>=0.0.79
insightface>=0.7.3
facenet-pytorch>=2.5.3
face-recognition>=1.3.0

# Natural Language Processing
spacy>=3.6.1
transformers>=4.30.0
nltk>=3.8.1
textblob>=0.17.1

# Machine Learning & Deep Learning
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
keras>=2.13.0

# Data Visualization & Analytics
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# OCR & Document Processing
pytesseract>=0.3.10
easyocr>=1.7.0
pdf2image>=1.16.0

# Audio Processing (for voice features)
speechrecognition>=3.10.0
pyaudio>=0.2.11
librosa>=0.10.0

# Database & Caching
redis>=4.6.0
sqlalchemy>=2.0.0
pymongo>=4.4.0

# API & Web Framework Extensions
flask-cors>=4.0.0
flask-socketio>=5.3.0
celery>=5.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Performance & Optimization
numba>=0.57.0
cython>=3.0.0
joblib>=1.3.0

# Security & Encryption
cryptography>=41.0.0
pyjwt>=2.8.0

# Monitoring & Logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Development Tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

